+++ data/champions.js
@@ -1,0 +1,348 @@
+// AVHL champions and championship rosters, 2022-23 through 2025-26.
+
+export const champions = [
+  {
+    "season": "2022-23",
+    "team": "Salt Lake City Scorpions",
+    "roster": [
+      {
+        "name": "Nicklas Backstrom",
+        "captain": false
+      },
+      {
+        "name": "T.J. Brodie",
+        "captain": false
+      },
+      {
+        "name": "Michael Bunting",
+        "captain": false
+      },
+      {
+        "name": "Dennis Cholowski",
+        "captain": false
+      },
+      {
+        "name": "Lawson Crouse",
+        "captain": false
+      },
+      {
+        "name": "Leon Draisaitl",
+        "captain": true
+      },
+      {
+        "name": "Chris Driedger",
+        "captain": false
+      },
+      {
+        "name": "Ryan Ellis",
+        "captain": false
+      },
+      {
+        "name": "Brett Howden",
+        "captain": false
+      },
+      {
+        "name": "Martin Jones",
+        "captain": false
+      },
+      {
+        "name": "Phil Kessel",
+        "captain": false
+      },
+      {
+        "name": "Travis Konecny",
+        "captain": false
+      },
+      {
+        "name": "Josh Manson",
+        "captain": false
+      },
+      {
+        "name": "Jared McIsaac",
+        "captain": false
+      },
+      {
+        "name": "Colin Miller",
+        "captain": false
+      },
+      {
+        "name": "T.J. Oshie",
+        "captain": false
+      },
+      {
+        "name": "Tanner Pearson",
+        "captain": false
+      },
+      {
+        "name": "Jason Robertson",
+        "captain": false
+      },
+      {
+        "name": "Ryan Strome",
+        "captain": false
+      },
+      {
+        "name": "Pius Suter",
+        "captain": false
+      }
+    ]
+  },
+  {
+    "season": "2023-24",
+    "team": "New York Steamrollers",
+    "roster": [
+      {
+        "name": "Pontus Aberg",
+        "captain": false
+      },
+      {
+        "name": "Samuel Bolduc",
+        "captain": false
+      },
+      {
+        "name": "Nick Bonino",
+        "captain": false
+      },
+      {
+        "name": "Logan Couture",
+        "captain": false
+      },
+      {
+        "name": "Jonathan Dahlén",
+        "captain": false
+      },
+      {
+        "name": "Magnus Hellberg",
+        "captain": false
+      },
+      {
+        "name": "Vinnie Hinostroza",
+        "captain": false
+      },
+      {
+        "name": "Jack Johnson",
+        "captain": false
+      },
+      {
+        "name": "Ross Johnston",
+        "captain": false
+      },
+      {
+        "name": "Chris Kreider",
+        "captain": false
+      },
+      {
+        "name": "Tyler Motte",
+        "captain": false
+      },
+      {
+        "name": "Stefan Noesen",
+        "captain": false
+      },
+      {
+        "name": "Jamie Oleksiak",
+        "captain": false
+      },
+      {
+        "name": "Max Pacioretty",
+        "captain": true
+      },
+      {
+        "name": "Jacob Perreault",
+        "captain": false
+      },
+      {
+        "name": "Calvin Pickard",
+        "captain": false
+      },
+      {
+        "name": "Filip Roos",
+        "captain": false
+      },
+      {
+        "name": "Justin Sourdif",
+        "captain": false
+      },
+      {
+        "name": "Shea Theodore",
+        "captain": false
+      },
+      {
+        "name": "Nikita Zadorov",
+        "captain": false
+      }
+    ]
+  },
+  {
+    "season": "2024-25",
+    "team": "Denver Mountain Lions",
+    "roster": [
+      {
+        "name": "Noel Acciari",
+        "captain": false
+      },
+      {
+        "name": "Jesper Boqvist",
+        "captain": false
+      },
+      {
+        "name": "Pavel Buchnevich",
+        "captain": false
+      },
+      {
+        "name": "Sidney Crosby",
+        "captain": true
+      },
+      {
+        "name": "Noah Dobson",
+        "captain": false
+      },
+      {
+        "name": "Robby Fabbri",
+        "captain": false
+      },
+      {
+        "name": "Brendan Gallagher",
+        "captain": false
+      },
+      {
+        "name": "Ridly Greig",
+        "captain": false
+      },
+      {
+        "name": "Kaiden Guhle",
+        "captain": false
+      },
+      {
+        "name": "Tyson Jost",
+        "captain": false
+      },
+      {
+        "name": "Philipp Kurashev",
+        "captain": false
+      },
+      {
+        "name": "Anders Lee",
+        "captain": false
+      },
+      {
+        "name": "Ryan Lindgren",
+        "captain": false
+      },
+      {
+        "name": "Nils Lundkvist",
+        "captain": false
+      },
+      {
+        "name": "John Marino",
+        "captain": false
+      },
+      {
+        "name": "William Nylander",
+        "captain": false
+      },
+      {
+        "name": "Tucker Poolman",
+        "captain": false
+      },
+      {
+        "name": "Brayden Schenn",
+        "captain": false
+      },
+      {
+        "name": "Arturs Silovs",
+        "captain": false
+      },
+      {
+        "name": "Anthony Stolarz",
+        "captain": false
+      }
+    ]
+  },
+  {
+    "season": "2025-26",
+    "team": "Cleveland Demons",
+    "roster": [
+      {
+        "name": "Mike Amadio",
+        "captain": false
+      },
+      {
+        "name": "Adam Boqvist",
+        "captain": false
+      },
+      {
+        "name": "Filip Chytil",
+        "captain": false
+      },
+      {
+        "name": "Matt Coronato",
+        "captain": false
+      },
+      {
+        "name": "Zach Dean",
+        "captain": false
+      },
+      {
+        "name": "Anthony Duclair",
+        "captain": false
+      },
+      {
+        "name": "William Eklund",
+        "captain": false
+      },
+      {
+        "name": "Radek Faksa",
+        "captain": false
+      },
+      {
+        "name": "Filip Forsberg",
+        "captain": false
+      },
+      {
+        "name": "Jordan Greenway",
+        "captain": false
+      },
+      {
+        "name": "Dougie Hamilton",
+        "captain": false
+      },
+      {
+        "name": "Connor Hellebuyck",
+        "captain": false
+      },
+      {
+        "name": "Alexis Lafrenière",
+        "captain": false
+      },
+      {
+        "name": "Brayden McNabb",
+        "captain": false
+      },
+      {
+        "name": "Kyle Palmieri",
+        "captain": false
+      },
+      {
+        "name": "Brayden Point",
+        "captain": true
+      },
+      {
+        "name": "Ilya Samsonov",
+        "captain": false
+      },
+      {
+        "name": "Jake Sanderson",
+        "captain": false
+      },
+      {
+        "name": "Yegor Zamula",
+        "captain": false
+      },
+      {
+        "name": "Calvin de Haan",
+        "captain": false
+      }
+    ]
+  }
+];